requirements.txt
flask
gunicorn
openai-whisper
ffmpeg-python
torch==2.0.0
numpy
